fix default repository, fix post rating, began to test comments

+++ Infrastructure/AutomapperProfiles/AutomapperPostProfile.cs
@@ -20,7 +20,7 @@ namespace Infrastructure.AutomapperProfiles
                 .ForMember(a => a.PostId, a => a.MapFrom(b => b.Id))
                 .ForMember(a => a.BriefDesctiption, a => a.MapFrom(b => b.BriefDesctiption))
                 .ForMember(a => a.Header, a => a.MapFrom(b => b.Header))
-                .ForMember(a => a.Score, a => a.MapFrom(b => (b.CountOfScore == 0) ? 0 : b.SumOfScore / b.CountOfScore))
+                .ForMember(a => a.Score, a => a.MapFrom(b => (b.CountOfScore == 0) ? 0 : (double)b.SumOfScore / b.CountOfScore))
                 .ForAllOtherMembers(a => a.Ignore());
             CreateMap<PostEntity, PostViewModel>()
                 .ForMember(a => a.PostId, a => a.MapFrom(b => b.Id))
@@ -29,7 +29,7 @@ namespace Infrastructure.AutomapperProfiles
                 .ForMember(a => a.Header, a => a.MapFrom(b => b.Header))
                 .ForMember(a => a.Section, a => a.MapFrom(b => (b.Section != null) ? b.Section.Name : null))
                 .ForMember(a => a.SectionId, a => a.MapFrom(b => b.SectionId))
-                .ForMember(a => a.Score, a => a.MapFrom(b => (b.CountOfScore == 0) ? 0 : b.SumOfScore / b.CountOfScore))
+                .ForMember(a => a.Score, a => a.MapFrom(b => (b.CountOfScore == 0) ? 0 : (double)b.SumOfScore / b.CountOfScore))
                 .ForAllOtherMembers(a => a.Ignore());
 
             CreateMap<PostCreateEditViewModel, PostEntity>()
